# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projeto de PDWA5",
        "description": "API do projeto de PDWA5",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Endpoint da API."
        }
    ],
    "paths": {
...
        "/menu/:permalink/block": {
            "get": {
                "tags": [
                    "Block"
                ],
                "summary": "Listar blocos.",
                "description": "Lista os blocos dentro dos menus.",
                "operationId": "b71f5c1063fd7c9467ddc254da6fe6b2",
                "responses": {
                    "200": {
                        "description": "Blocos lisitados com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "user": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "menu": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "text": {
                                                "type": "string",
                                                "example": "Novo Bloco"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2022-11-14T00:06:03.000000Z"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2022-11-14T00:06:03.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ao falhar na autorização."
                    }
                }
            },
            "post": {
                "tags": [
                    "Block"
                ],
                "summary": "Criar bloco.",
                "description": "Cria um novo bloco.",
                "operationId": "5c5b2997ae6d85d19c4ace9047fe0421",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Novo Bloco"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "user": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "menu": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "text": {
                                            "type": "string",
                                            "example": "Novo Bloco"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2022-11-14T00:06:03.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2022-11-14T00:06:03.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ao falhar na autorização."
                    },
                    "422": {
                        "description": "Ao falhar em uma validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "text": {
                                                "type": "array",
                                                "items": {}
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/menu/:permalink/block/:id": {
            "put": {
                "tags": [
                    "Block"
                ],
                "summary": "Editar usuário.",
                "description": "Edita um usuário.",
                "operationId": "c3b0d055445303ff3dd60db1d0d22ad9",
                "parameters": [
                    {
                        "name": "permalink",
                        "in": "path",
                        "description": "Permalink do menu",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do bloco",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Novo Bloco"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "user": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "menu": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "text": {
                                            "type": "string",
                                            "example": "Novo Bloco"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2022-11-14T00:06:03.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2022-11-14T00:06:03.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ao falhar na autorização."
                    },
                    "409": {
                        "description": "Caso já exista um menu com o mesmo nome."
                    },
                    "422": {
                        "description": "Ao falhar em uma validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "text": {
                                                "type": "array",
                                                "items": {}
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Block"
                ],
                "summary": "Remover bloco.",
                "description": "Remove o bloco.",
                "operationId": "1bbd38106aaa5527f8abc1ae5142add3",
                "parameters": [
                    {
                        "name": "permalink",
                        "in": "path",
                        "description": "Permalink do menu",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do bloco",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bloco removido com sucesso."
                    },
                    "401": {
                        "description": "Ao falhar na autorização."
                    },
                    "404": {
                        "description": "Caso o menu não exista."
                    }
                }
            }
        },
...
    }
}